fsd architecture implemented, file system restructured, absolute paths implemented in tsconfig and webpack
