+++ config/webpack/defineResolution.ts
@@ -1,7 +1,11 @@
 import webpack from "webpack";
 
-export const defineResolution = (): webpack.ResolveOptions => {
+export const defineResolution = (sourcePath: string): webpack.ResolveOptions => {
     return {
-        extensions: [".ts", ".tsx", ".js", ".jsx"]
+        extensions: [".ts", ".tsx", ".js", ".jsx"],
+        preferAbsolute: true,
+        modules: [sourcePath, "node_modules"],
+        alias: {},
+        mainFiles: ["index"]
     }
 };
