+++ config/webpack/interfaces/IWebpackConfigPaths.ts
@@ -2,4 +2,5 @@ export interface IWebpackConfigPaths {
     entry: string,
     output: string,
     template: string,
+    source: string
 }
