--- src/app/pages/index.tsx
@@ -1,3 +1,0 @@
-export { AboutPageLazy } from "./AboutPage/AboutPage.lazy";
-export { MainPageLazy } from "./MainPage/MainPage.lazy";
-export { CounterPageLazy } from "./CounterPage/CounterPage.lazy";
