+++ src/app/styles/styles.global/index.scss
@@ -1,6 +1,6 @@
 @import "./variables/variables.scss";
-@import "./theme/light.scss";
-@import "./theme/dark.scss";
+@import "../../themes/styles/styles.themes/light.scss";
+@import "../../themes/styles/styles.themes/dark.scss";
 @import "./reset.scss";
 
 body {
