+++ src/app/styles/styles.global/variables/_variables.scss
