+++ src/app/themes/lib/types/ETheme.ts
