+++ src/app/themes/lib/useTheme.tsx
@@ -1,0 +1,7 @@
+import { createContext, useContext } from "react";
+import { ITheme } from "./interfaces/ITheme";
+
+export const ThemeContext = createContext<ITheme>({});
+
+export const useTheme = () => useContext(ThemeContext);
+
