+++ src/app/themes/styles/styles.themes/dark.scss
