+++ src/app/themes/ui/ThemeProvider.tsx
@@ -1,21 +1,17 @@
-import { localStorageService } from "../../services/localStorageService/localStorageService";
-import { createContext, useContext, useState, useMemo, FC } from "react";
-import { LS_THEME_KEY } from "../../themes/constants/LSThemeKey";
-import { ITheme } from "../../themes/interfaces/ITheme";
-import { ETheme } from "../../themes/types/ETheme";
-import { LS_APP_KEY } from "../LSAppKeys";
+import { localStorageService } from "../../../shared/services/localStorageService/localStorageService";
+import { LS_THEME_KEY } from "../lib/constants/LSThemeKey";
+import React, { useState, useMemo, FC } from "react";
+import { ThemeContext } from "../lib/useTheme";
+import { ETheme } from "../lib/types/ETheme";
+import { LS_APP_KEY } from "../../LSAppKeys";
 
 const initialState = localStorageService.read(LS_APP_KEY)?.[LS_THEME_KEY] as ETheme ?? ETheme.LIGHT;
-
-const ThemeContext = createContext<ITheme>({});
-
-export const useTheme = () => useContext(ThemeContext);
 
 export const ThemeProvider: FC = ({ children }) => {
     const [theme, setTheme] = useState<ETheme>(initialState);
 
     const themeSwitchHandler = (): void => {
-        setTheme(prevTheme => {
+        setTheme((prevTheme: string) => {
             const switchedTheme = prevTheme === ETheme.DARK ? ETheme.LIGHT : ETheme.DARK;
             localStorageService.create(LS_APP_KEY, {[LS_THEME_KEY]: switchedTheme});
             return switchedTheme;
