+++ src/pages/AboutPage/AboutPage.lazy.tsx
