+++ src/pages/CounterPage/CounterPage.lazy.tsx
