+++ src/pages/MainPage/MainPage.lazy.tsx
