+++ src/pages/index.tsx
@@ -1,0 +1,3 @@
+export { AboutPageLazy as AboutPage } from "./AboutPage/AboutPage.lazy";
+export { MainPageLazy as MainPage } from "./MainPage/MainPage.lazy";
+export { CounterPageLazy as CounterPage } from "./CounterPage/CounterPage.lazy";
