+++ src/pages/styles/styles.pages/Counter.module.scss
