+++ src/shared/components/Button/Button.tsx
@@ -1,4 +1,4 @@
-import classes from "../../../styles/styles.components/Button.module.scss";
+import classes from "../styles/styles.components/Button.module.scss";
 import { FC } from "react";
 
 export const Button: FC<{title: string, onClick: () => void}> = ({ title, onClick }) => {
