+++ src/shared/components/styles/styles.components/Loader.module.scss
