+++ tsconfig.json
@@ -2,6 +2,10 @@
     "compilerOptions": {
         "outDir": "dist",
         "sourceMap": true,
+        "baseUrl": ".",
+        "paths": {
+            "*": ["./src/*"]
+        },
         "noImplicitAny": true,
         "module": "ESNext",
         "target": "es5",
